requirements.txt
yfinance
pandas
numpy
scikit-learn
xgboost
lightgbm
ccxt
requests
beautifulsoup4
ta
plotly
dash
python-dotenv
schedule
transformers
torch
feedparser
alpaca-trade-api
flask
gunicorn